Pull out part of arena views into new props component
