+++ .vscode/settings.json
@@ -6,12 +6,15 @@
         "fullscreen",
         "fullscreen type",
         "gameplay",
+        "hittester",
+        "infobar",
         "keymod",
         "knockback",
         "lifebar",
         "rgba",
         "rustc",
         "sdl",
+        "skillbar",
         "type"
     ]
 }
