+++ docs/module_map.txt
@@ -23,22 +23,22 @@ Between Combat UI    |  intermission   +----+   arena    |   Combat UI
                              |           |                           |
                              +-----+-----+                           |
                                    |                                 |
-                                   |      No components / state      |
-                                   |           below here            |
-                                   |                                 |
-                                   |                           +-----+---------+
-                                   |                           |               |
-                                   |                           |   conductor   |
-                                   |                           |               |
-                                   |                           +-----+---------+
-                                   |                                 |
+                                   |No components / state            |
+                                   |     below here     +---------------------------+
+                                   |                    |            |              |
+                                   |            +-------+--------    |      +-------+-------+
+                                   |            |               |    |      |               |
+                                   |            |    props      |    |      |   conductor   |
+                                   |            |               |    |      |               |
+                                   |            +------+--------+    |      +--------+------+
+                                   |                   |                             |
+                                   |                   +-------------+---------------+
                                    |                                 |
                                    |                           +-----+-----------+
                                    |                           |                 |
                                    |                           |   after_image   |
                                    |                           |                 |
                                    |                           +-----+-----------+
-                                   |                                 |
                                    |                                 |
                                    +------------+--------------------+
                                                 |
