+++ src/arena/views.rs
@@ -1,9 +1,3 @@
-use specs::prelude::*;
-
-use crate::after_image::{LayoutChunkIcon, RenderCanvas};
-use crate::atlas::prelude::*;
-use crate::clash::StatusKind;
-
 mod character_overlay;
 mod debug;
 mod help_popup;
@@ -14,32 +8,6 @@ mod skillbar;
 mod status_display;
 mod text_hittester;
 mod text_render_helper;
-pub mod view_components;
-
-#[allow(dead_code)]
-#[derive(is_enum_variant, Clone, Debug)]
-pub enum HitTestResult {
-    None,
-    Skill(String),
-    Tile(Point),
-    Enemy(Point),
-    Field(Point),
-    Orb(Point),
-    Icon(LayoutChunkIcon),
-    Text(String),
-    Status(StatusKind),
-    BackButton,
-    CloseButton,
-    Button,
-    // Consider adding a check in lookup_hittest if we need it to display help
-}
-
-pub trait View {
-    fn render(&self, ecs: &World, canvas: &mut RenderCanvas, frame: u64) -> BoxResult<()>;
-    fn hit_test(&self, _ecs: &World, _x: i32, _y: i32) -> Option<HitTestResult> {
-        None
-    }
-}
 
 pub use character_overlay::CharacterOverlay;
 pub use debug::DebugView;
