+++ src/props/views.rs
@@ -7,9 +7,9 @@ use sdl2::rect::Rect as SDLRect;
 use sdl2::render::Texture;
 use specs::prelude::*;
 
-use super::{HitTestResult, View};
 use crate::after_image::prelude::*;
 use crate::atlas::prelude::*;
+use crate::props::{HitTestResult, View};
 
 pub struct Frame {
     position: SDLPoint,
